Separated ViewController and ViewModel in PersonDetailViewController file.
